+++ src/constructors/constructWithOptions.js
@@ -1,5 +1,6 @@
 // @flow
-import type { Interpolation, Target } from '../types'
+import { isValidElementType } from 'react-is'
+import type { Target } from '../types'
 
 export default (css: Function) => {
   const constructWithOptions = (
@@ -7,7 +8,7 @@ export default (css: Function) => {
     tag: Target,
     options: Object = {}
   ) => {
-    if (typeof tag !== 'string' && typeof tag !== 'function') {
+    if (!isValidElementType(tag)) {
       throw new Error(
         process.env.NODE_ENV !== 'production'
           ? `Cannot create styled-component for component: ${String(tag)}`
@@ -16,10 +17,9 @@ export default (css: Function) => {
     }
 
     /* This is callable directly as a template function */
-    const templateFunction = (
-      strings: Array<string>,
-      ...interpolations: Array<Interpolation>
-    ) => componentConstructor(tag, options, css(strings, ...interpolations))
+    // $FlowFixMe: Not typed to avoid destructuring arguments
+    const templateFunction = (...args) =>
+      componentConstructor(tag, options, css(...args))
 
     /* If config methods are called, wrap up a new template function and merge options */
     templateFunction.withConfig = config =>