+++ src/constructors/injectGlobal.js
@@ -3,18 +3,21 @@ import hashStr from '../vendor/glamor/hash'
 import StyleSheet from '../models/StyleSheet'
 import type { Interpolation, Stringifier } from '../types'
 
+type InjectGlobalFn = (
+  strings: Array<string>,
+  ...interpolations: Array<Interpolation>
+) => void
+
 export default (stringifyRules: Stringifier, css: Function) => {
-  const injectGlobal = (
-    strings: Array<string>,
-    ...interpolations: Array<Interpolation>
-  ) => {
-    const rules = css(strings, ...interpolations)
+  const injectGlobal: InjectGlobalFn = (...args) => {
+    const styleSheet = StyleSheet.master
+    const rules = css(...args)
     const hash = hashStr(JSON.stringify(rules))
+    const id = `sc-global-${hash}`
 
-    const componentId = `sc-global-${hash}`
-    if (StyleSheet.instance.hasInjectedComponent(componentId)) return
-
-    StyleSheet.instance.inject(componentId, false, stringifyRules(rules))
+    if (!styleSheet.hasId(id)) {
+      styleSheet.inject(id, stringifyRules(rules))
+    }
   }
 
   return injectGlobal