+++ src/test/utils.js
@@ -18,9 +18,15 @@ import noParserStringifyRules from '../no-parser/stringifyRules'
 
 /* Ignore hashing, just return class names sequentially as .a .b .c etc */
 let index = 0
+let inputs = {}
 let seededClassnames = []
-const classNames = () =>
-  seededClassnames.shift() || String.fromCodePoint(97 + index++)
+
+const classNames = input => {
+  const seed = seededClassnames.shift()
+  if (seed) return seed
+
+  return inputs[input] || (inputs[input] = String.fromCodePoint(97 + index++))
+}
 
 export const seedNextClassnames = (names: Array<string>) =>
   (seededClassnames = names)
@@ -36,6 +42,7 @@ export const resetStyled = (isServer: boolean = false) => {
 
   StyleSheet.reset(isServer)
   index = 0
+  inputs = {}
 
   const ComponentStyle = _ComponentStyle(classNames, flatten, stringifyRules)
   const constructWithOptions = _constructWithOptions(css)
@@ -78,12 +85,16 @@ export const expectCSSMatches = (
   opts: { ignoreWhitespace: boolean } = { ignoreWhitespace: true }
 ) => {
   // NOTE: This should normalise both CSS strings to make irrelevant mismatches less likely
-  const expectation = _expectation.replace(/ {/g, '{').replace(/:\s+;/g, ':;')
+  const expectation = _expectation
+    .replace(/ {/g, '{')
+    .replace(/:\s+/g, ':')
+    .replace(/:\s+;/g, ':;')
 
   const css = Array.from(document.querySelectorAll('style'))
     .map(tag => tag.innerHTML)
     .join('\n')
     .replace(/ {/g, '{')
+    .replace(/:\s+/g, ':')
     .replace(/:\s+;/g, ':;')
 
   if (opts.ignoreWhitespace) {