+++ src/utils/stringifyRules.js
@@ -3,13 +3,23 @@ import Stylis from 'stylis'
 import _insertRulePlugin from 'stylis-rule-sheet'
 import type { Interpolation } from '../types'
 
+// NOTE: This stylis instance is only used to split rules from SSR'd style tags
+const stylisSplitter = new Stylis({
+  global: false,
+  cascade: false,
+  keyframe: false,
+  prefix: false,
+  compress: false,
+  semicolon: true,
+})
+
 const stylis = new Stylis({
   global: false,
   cascade: true,
   keyframe: false,
   prefix: true,
   compress: false,
-  semicolon: true,
+  semicolon: false, // NOTE: This means "autocomplete missing semicolons"
 })
 
 // Wrap `insertRulePlugin to build a list of rules,
@@ -31,6 +41,7 @@ const parseRulesPlugin = _insertRulePlugin(rule => {
 })
 
 stylis.use([parseRulesPlugin, returnRulesPlugin])
+stylisSplitter.use([parseRulesPlugin, returnRulesPlugin])
 
 const stringifyRules = (
   rules: Array<Interpolation>,
@@ -44,5 +55,8 @@ const stringifyRules = (
 
   return stylis(prefix || !selector ? '' : selector, cssStr)
 }
+
+export const splitByRules = (css: string): Array<string> =>
+  stylisSplitter('', css)
 
 export default stringifyRules