+++ typings/styled-components.d.ts
@@ -16,9 +16,10 @@ export type ThemedOuterStyledProps<P, T> = P & {
 };
 export type OuterStyledProps<P> = ThemedOuterStyledProps<P, any>;
 
+export type FalseyValue = undefined | null | false;
 export type Interpolation<P> = FlattenInterpolation<P> | ReadonlyArray<FlattenInterpolation<P> | ReadonlyArray<FlattenInterpolation<P>>>;
 export type FlattenInterpolation<P> = InterpolationValue | InterpolationFunction<P>;
-export type InterpolationValue = string | number | Styles | StyledComponentClass<any, any>;
+export type InterpolationValue = string | number | Styles | FalseyValue | StyledComponentClass<any, any>;
 export type SimpleInterpolation = InterpolationValue | ReadonlyArray<InterpolationValue | ReadonlyArray<InterpolationValue>>;
 export interface Styles {
   [ruleOrSelector: string]: string | number | Styles;
@@ -93,6 +94,7 @@ export function withTheme<P extends { theme?: T; }, T>(component: Component<P>):
 
 export function keyframes(strings: TemplateStringsArray, ...interpolations: SimpleInterpolation[]): string;
 export function injectGlobal(strings: TemplateStringsArray, ...interpolations: SimpleInterpolation[]): void;
+export function consolidateStreamedStyles(): void;
 
 export const ThemeProvider: ThemeProviderComponent<object>;
 
@@ -100,13 +102,20 @@ interface StylesheetComponentProps {
   sheet: ServerStyleSheet;
 }
 
-export class StyleSheetManager extends React.Component<StylesheetComponentProps, {}> { }
+interface StyleSheetManagerProps {
+  sheet?: StyleSheet;
+  target?: Node;
+}
+
+export class StyleSheetManager extends React.Component<StyleSheetManagerProps, {}> { }
 
 export class ServerStyleSheet {
   collectStyles(tree: React.ReactNode): ReactElement<StylesheetComponentProps>;
 
   getStyleTags(): string;
   getStyleElement(): ReactElement<{}>[];
+  interleaveWithNodeStream(readableStream: NodeJS.ReadableStream): NodeJS.ReadableStream;
+  instance: StyleSheet;
 }
 
 export default styled;